Add gc use intrinisic
Now that our gc root placement pass is significantly more aggressive about
dropping roots, we need to be a bit more careful about our use of pointer,
etc. This adds a low-level intrinsic for annotating gc uses to keep objects
alive even if they would be otherwise unreferenced.

As an initial use case, we get rid of a number of uses of `pointer` in string,
but creating a new `unsafe_load` that keeps the string alive.